Primeras predicciones con LDA y KNN k=3 sklearn

+++ preprocessing/datasets_preparation.py
@@ -28,13 +28,13 @@ def dataset_prep():
     if os.path.exists(filename):
         df_rd = load(filename,ct_rd,mode)
         df_rd_et = load(filename,et_rd,mode)
-        train_rd = np.vstack((df_rd,df_rd_et)).astype(np.float16)
+        train_rd = np.vstack((df_rd,df_rd_et))
         save(filename_ds, 'train_rd', train_rd)
         df_r = load(filename,ct_r,mode)
         df_r_et = load(filename,et_r,mode)
-        train_r = np.vstack((df_r,df_r_et)).astype(np.float16)
+        train_r = np.vstack((df_r,df_r_et))
         save(filename_ds, 'train_r', train_r)
-        labels = np.hstack((ct_labels,et_labels)).astype(np.int8)
+        labels = np.hstack((ct_labels,et_labels))
         save(filename_ds, 'labels', labels)
         
 dataset_prep()
